Shoot Fireball
The wizard now shoots a fireball in the direction that he is facing when the q key is pressed during game play.

this is for task #77

+++ Wizard Hack/Assets/Scripts/PlayerHealth.cs
@@ -8,8 +8,9 @@ public class PlayerHealth : MonoBehaviour
 
     public int initialHealth = 100;
     public int playerCurrentHealth;
-    public int initialMana = 100;
-    public int playerCurrentMana;
+    public float initialMana = 100;
+    public float playerCurrentMana;
+   
     public Slider playerHealthSlider;
     public Slider playerManaSlider;
     public Image damageImage;
@@ -27,7 +28,7 @@ public class PlayerHealth : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        
+       
     }
 
 
@@ -43,13 +44,13 @@ public class PlayerHealth : MonoBehaviour
         playerHealthSlider.value = playerCurrentHealth;
     }
 
-    public void useMana(int usedAmount)
+    public void useMana(float usedAmount)
     {
         playerCurrentMana -= usedAmount;
         playerManaSlider.value = playerCurrentMana;
     }
 
-    public void gainMana(int gainAmount)
+    public void gainMana(float gainAmount)
     {
         playerCurrentMana += gainAmount;
         playerManaSlider.value = playerCurrentMana;
